requirements.txt
python-telegram-bot==21.7
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
fake-useragent==1.4.0
aiohttp==3.9.3
ratelimit==2.2.1
pytest==8.0.0
pytest-asyncio>=0.23.5
pytest-cov==4.1.0
